+++ packages/chain-interfaces/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @phantom/chain-interfaces
 
+## 1.0.4
+
+### Patch Changes
+
+- 2d00fb0: Do not throw error if one of the chains is not available on connection
+- Updated dependencies [2d00fb0]
+  - @phantom/constants@1.0.4
+  - @phantom/parsers@1.0.4
+  - @phantom/sdk-types@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes