+++ packages/constants/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @phantom/constants
 
+## 1.0.4
+
+### Patch Changes
+
+- 2d00fb0: Do not throw error if one of the chains is not available on connection
+- Updated dependencies [2d00fb0]
+  - @phantom/sdk-types@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes